+++ WireData/Package.swift
@@ -4,6 +4,7 @@ import PackageDescription
 
 let package = Package(
     name: "WireData",
+    platforms: [.iOS(.v17)],
     products: [
         .library(
             name: "WireData",
