+++ WireData/Tests/TestPlans/AllTests.xctestplan
@@ -9,7 +9,19 @@
     }
   ],
   "defaultOptions" : {
-    "codeCoverage" : false
+    "commandLineArgumentEntries" : [
+      {
+        "argument" : "-com.apple.CoreData.ConcurrencyDebug 1"
+      }
+    ],
+    "environmentVariableEntries" : [
+      {
+        "key" : "CI",
+        "value" : "${CI}"
+      }
+    ],
+    "language" : "en",
+    "region" : "DE"
   },
   "testTargets" : [
     {
