+++ WireData/Tests/WireDataTests/Utilities/NSPersistentContainer+inMemoryContainer.swift
@@ -16,24 +16,26 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
+import CoreData
 import WireData
-import WireDataModel
 
-public struct NewUserInfo: Equatable, Sendable {
-    let userID: WireDataModel.QualifiedID
-    let name: String
-    let handle: String?
-    let teamID: UUID?
-    let type: WireDataModel.TypeOfUser?
-    let accentID: Int
-    let previewAssetKey: String?
-    let completeAssetKey: String?
-    let isDeleted: Bool
-    let email: String?
-    let expiresAt: Date?
-    let appDescription: String?
-    let appCategory: String?
-    let serviceID: UUID?
-    let serviceProvider: UUID?
-    let supportedProtocols: Set<WireDataModel.MessageProtocol>?
+extension NSPersistentContainer {
+
+    static func inMemoryContainer() throws -> NSPersistentContainer {
+        let model = NSManagedObjectModel.mergedModel(from: [WireDataBundle.bundle])
+
+        let description = NSPersistentStoreDescription()
+        description.type = NSInMemoryStoreType
+        let container = NSPersistentContainer(name: "zmessaging", managedObjectModel: model!)
+        container.persistentStoreDescriptions = [description]
+        var loadError: Error?
+        container.loadPersistentStores { _, error in
+            loadError = error
+        }
+        if let loadError {
+            throw loadError
+        }
+        return container
+    }
+
 }
